+++ docs-src/components/index.js
@@ -1,27 +1,33 @@
 export { Autocomplete } from './autocomplete';
 export { Button } from './button';
+export { PushButton } from './push-button';
 export { ButtonGroup } from './button-group';
 export { ButtonToggle } from './button-toggle';
-export { Checkbox } from './checkbox';
-export { ColorPalette } from './color-palette';
 export { Datepicker } from './datepicker';
+
+export { Checkbox } from './checkbox';
+export { InputMath } from './input-math';
+export { InputNumber } from './input-number';
+export { InputPassword } from './input-password';
+export { InputText } from './input-text';
+export { Radio } from './radio';
+export { Select } from './select';
+export { Textarea } from './textarea';
+export { Toggle } from './toggle';
+
 export { Dialog } from './dialog';
 export { Drawer } from './drawer';
-export { Icon } from './icon';
-export { Input } from './input';
-export { InputMath } from './input-math';
-export { InputPassword } from './input-password';
+export { InfoBar } from './info-bar';
 export { Menu } from './menu';
 export { MessageBox } from './message-box';
 export { Panel } from './panel';
-export { PushButton } from './push-button';
-export { Select } from './select';
 export { Splitter } from './splitter';
 export { Table } from './table';
-export { TextFit } from './text-fit';
-export { Textarea } from './textarea';
-export { Toaster } from './toaster';
-export { NotificationCenter } from './notification-center';
-export { Toggle } from './toggle';
-export { Tooltip } from './tooltip';
 export { Tree } from './tree';
+
+export { Icon } from './icon';
+export { NotificationCenter } from './notification-center';
+export { TextFit } from './text-fit';
+export { Toaster } from './toaster';
+export { Tooltip } from './tooltip';
+export { ColorPalette } from './color-palette';