+++ docs-src/nav/Nav.svelte
@@ -24,11 +24,14 @@
 		<NavItem name="Push Button" {active} />
 		<NavItem name="Button Group" {active} />
 		<NavItem name="Button Toggle" {active} />
+
 		<NavItem name="Checkbox" {active} />
 		<NavItem name="Datepicker" {active} />
-		<NavItem name="Input" {active} />
+		<NavItem name="Input Text" {active} />
+		<NavItem name="Input Number" {active} />
 		<NavItem name="Input Math" {active} />
-		<NavItem name="Password" hash="InputPassword" {active} />
+		<NavItem name="Input Password" {active} />
+		<NavItem name="Radio" {active} />
 		<NavItem name="Select" {active} />
 		<NavItem name="Textarea" {active} />
 		<NavItem name="Toggle" {active} />
@@ -36,6 +39,7 @@
 		<h3>Containers</h3>
 		<NavItem name="Dialog" {active} />
 		<NavItem name="Drawer" {active} />
+		<NavItem name="InfoBar" {active} />
 		<NavItem name="Menu" {active} />
 		<NavItem name="MessageBox" {active} />
 		<NavItem name="Panel" {active} />