+++ docs-src/pages/changelog.svelte
@@ -1,4 +1,34 @@
 <h1>Changelog</h1>
+<h2>v7.0.0 <em>(2023-?)</em></h2>
+<ul>
+<li><strong>New:</strong> <a href="#InfoBar">InfoBar</a> component.</li>
+<li><strong>New:</strong> <a href="#InputText">InputText</a>, <a href="#InputNumber">InputNumber</a>, and <a href="#Radio">Radio</a> components.</li>
+<li><strong>New:</strong> <code>info</code>, <code>error</code> and <code>label</code> attributes are now supported in all basic inputs (<code>InputText</code>, <code>InputNumber</code>, <code>InputMath</code>, <code>InputPassword</code>, <code>Radio</code>, and <code>Checkbox</code>).</li>
+<li><strong>Improved:</strong> <code>InputMath</code> component: support for <code>()</code> characters, to allow for more complex expressions.</li>
+</ul>
+<h3>Breaking changes</h3>
+<h5>Checkbox</h5>
+<ul>
+<li>HTML structure changed <code>input</code> --&gt; <code>.checkbox .checkbox-row input</code></li>
+<li><code>on:change</code> is called with a svelte event instead of the native one, so: <code>e.target.checked</code> is now <code>e.detail.checked</code></li>
+</ul>
+<h5>InputMath</h5>
+<ul>
+<li>HTML structure changed <code>.input-math-wrapper input</code> --&gt; <code>.input-math .input-text-inner .input-math-row input</code></li>
+</ul>
+<h5>InputNumber:</h5>
+<ul>
+<li>HTML structure changed: <code>input</code> --&gt; <code>.input-number .input-text-inner input</code></li>
+</ul>
+<h5>InputPassword</h5>
+<ul>
+<li>HTML structure changed: <code>.input-password-wrapper .input-password-row input</code> --&gt; <code>.input-password .input-text-inner .input-password-row input</code></li>
+</ul>
+<h5>CSS variables changed:</h5>
+<ul>
+<li><code>--ui-shadow-invalid</code> --&gt; <code>--ui-shadow-danger</code></li>
+</ul>
+<hr>
 <h2>v6.8.1, v6.8.2 <em>(2023-06-21)</em></h2>
 <ul>
 <li>Allow HTML in <code>MessageBox</code>.</li>