+++ src/info-bar/index.js
@@ -1,0 +1,4 @@
+export { default as Error } from './Error.svelte';
+export { default as Info } from './Info.svelte';
+export { default as Warning } from './Warning.svelte';
+export { default as InfoBar } from './InfoBar.svelte';