+++ src/input/index.js
@@ -1,0 +1,6 @@
+export * from './checkbox';
+export * from './input-math';
+export * from './input-number';
+export * from './input-password';
+export * from './input-text';
+export * from './radio';