+++ src/utils.js
@@ -114,7 +114,7 @@ export function innerHeight (el) {
 	return el.getBoundingClientRect().height - borders - padding;
 }
 
-export function uuid () {
+export function guid () {
 	return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, c => {
 		const r = Math.random() * 16 | 0, v = c === 'x' ? r : (r & 0x3 | 0x8);
 		return v.toString(16);