+++ tests/InputMath.spec.js
@@ -1,6 +1,7 @@
 import { render } from '@testing-library/svelte';
-import { InputMath } from '../src/input-math';
+import { InputMath } from '../src/input';
 import userEvent from '@testing-library/user-event';
+import { waitForTimeout } from './helpers/utils';
 
 
 const props = {
@@ -10,6 +11,8 @@ const props = {
 	placeholder: 'Component1',
 	class: 'test-class',
 	required: true,
+	error: 'error',
+	label: 'Component1',
 };
 
 
@@ -18,17 +21,44 @@ test('InputMath', async () => {
 	const mock = jest.fn();
 	component.$on('change', mock);
 
-	const inputMath = container.querySelector('.test-class');
-	expect(inputMath).toBeInTheDocument();
-	expect(inputMath).toHaveClass('test-class');
+	const cmp = container.querySelector('.test-class');
+	expect(cmp).toBeInTheDocument();
+	expect(cmp).toHaveClass('test-class');
 
 	// verify props
-	const input = getByTitle('Component1');
-	expect(input).toHaveAttribute('id', 'Component1');
-	expect(input).toHaveAttribute('title', 'Component1');
-	expect(input).toHaveAttribute('name', 'Component1');
-	expect(input).toHaveAttribute('placeholder', 'Component1');
-	expect(input).toHaveAttribute('required');
+	const input = getByTitle(props.title);
+	expect(input).toHaveAttribute('id', props.id);
+	expect(input).toHaveAttribute('title', props.title);
+	expect(input).toHaveAttribute('name', props.name);
+	expect(input).toHaveAttribute('placeholder', props.placeholder);
+	expect(input).toHaveAttribute('aria-required');
+
+	let err = cmp.querySelector('.info-bar-error');
+	expect(err).toBeInTheDocument();
+	expect(err).toHaveTextContent(props.error);
+
+	await component.$set({ error: '' });
+	await waitForTimeout();
+	err = cmp.querySelector('.info-bar-error');
+	expect(err).not.toBeInTheDocument();
+
+	await component.$set({ info: 'info' });
+	let info = cmp.querySelector('.info-bar-info');
+	expect(info).toBeInTheDocument();
+	expect(info).toHaveTextContent('info');
+
+	await component.$set({ info: '' });
+	await waitForTimeout();
+	info = cmp.querySelector('.info-bar-info');
+	expect(info).not.toBeInTheDocument();
+
+	const lbl = cmp.querySelector('label');
+	expect(lbl).toBeInTheDocument();
+	expect(lbl).toHaveAttribute('for', props.id);
+	expect(lbl).toHaveTextContent(props.label);
+
+
+
 
 	// test just digits
 	await userEvent.clear(input);